requirements.txt
Django==5.2
djangorestframework==3.14.0
python-dotenv==1.1.1
requests==2.31.0
tzdata==2024.1
